循環参照回避のため、modelsパッケージにモデル定義を移す
