+++ internal/models/record.go
@@ -1,4 +1,4 @@
-package internal
+package models
 
 import (
 	"time"
