+++ internal/models/user.go
@@ -1,10 +1,8 @@
-package internal
+package models
 
 import (
-	"net/http"
 	"time"
 
-	"github.com/labstack/echo/v4"
 	"github.com/uptrace/bun"
 )
 
@@ -20,22 +18,4 @@ type User struct {
 	UpdatedAt time.Time `json:"-" bun:",nullzero,notnull,default:current_timestamp"`
 
 	Subjects []*Subject `json:"-" bun:"rel:has-many,join:id=user_id"`
-}
-
-func (h *Handler) GetUser(c echo.Context) error {
-	p := c.Param("username")
-
-	var u User
-	err := h.DB.NewSelect().
-		Model(&u).
-		Where("username = ?", p).
-		Scan(c.Request().Context())
-	if err != nil {
-		return err
-	}
-
-	if err := c.JSON(http.StatusOK, u); err != nil {
-		return err
-	}
-	return nil
 }
